requirement.txt
pyyaml
xarray
netCDF4
numpy
matplotlib
torch